+++ integration/main_test.go
@@ -215,7 +215,7 @@ func InitializeScenario(s *godog.ScenarioContext) {
 	s.Step(`^the oracles broadcast data signed with "([^"]*)":$`, func(pubKeys string, properties *godog.Table) error {
 		return steps.OraclesBroadcastDataSignedWithKeys(execsetup.oracleEngine, pubKeys, properties)
 	})
-	s.Step(`^the followin LP events should be emitted:$`, func(table *godog.Table) error {
+	s.Step(`^the following LP events should be emitted:$`, func(table *godog.Table) error {
 		return steps.TheFollowingLPEventsShouldBeEmitted(execsetup.broker, table)
 	})