+++ integration/steps/lp_events.go
@@ -15,7 +15,7 @@ func TheFollowingLPEventsShouldBeEmitted(broker *stubs.BrokerStub, table *godog.
 	evtsByPartyID := map[string]map[string][]events.LiquidityProvision{}
 	for _, e := range lpEvts {
 		party := e.PartyID()
-		lpID := e.LiquidityProvision().Id
+		lpID := e.LiquidityProvision().Reference
 		m, ok := evtsByPartyID[party]
 		if !ok {
 			m = map[string][]events.LiquidityProvision{}
@@ -25,14 +25,13 @@ func TheFollowingLPEventsShouldBeEmitted(broker *stubs.BrokerStub, table *godog.
 			s = []events.LiquidityProvision{}
 		}
 		m[lpID] = append(s, e)
-		evtsByParty[party] = m
+		evtsByPartyID[party] = m
 	}
 	for _, row := range parseLPEventTable(table) {
 		lpe := LPEventWrapper{
 			row: row,
 		}
 		party, id, version, final, amt := lpe.Party(), lpe.ID(), lpe.Version(), lpe.Final(), lpe.CommitmentAmount()
-		ref, okRef := lpe.Reference()
 		lpIDMap, ok := evtsByPartyID[party]
 		if !ok {
 			return fmt.Errorf("no LP events found for party %s", party)
@@ -42,7 +41,7 @@ func TheFollowingLPEventsShouldBeEmitted(broker *stubs.BrokerStub, table *godog.
 			return fmt.Errorf("no LP events found for LP ID %s (party %s)", id, party)
 		}
 		if final {
-			if err := checkLPEvent(evts[len(evts)-1], party, id, ref, amt, version, okRef); err != nil {
+			if err := checkLPEvent(evts[len(evts)-1], party, id, amt, version); err != nil {
 				return err
 			}
 			continue
@@ -51,7 +50,7 @@ func TheFollowingLPEventsShouldBeEmitted(broker *stubs.BrokerStub, table *godog.
 		var err error
 		// find matching event
 		for _, e := range evts {
-			if err = checkLPEvent(e, party, id, ref, amt, version, okRef); err == nil {
+			if err = checkLPEvent(e, party, id, amt, version); err == nil {
 				// match found, break
 				break
 			}
@@ -63,14 +62,9 @@ func TheFollowingLPEventsShouldBeEmitted(broker *stubs.BrokerStub, table *godog.
 	return nil
 }
 
-func checkLPEvent(last events.LiquidityProvision, party, id, ref string, amt *num.Uint, version uint64, okRef bool) error {
+func checkLPEvent(last events.LiquidityProvision, party, id string, amt *num.Uint, version uint64) error {
 	if last.LiquidityProvision().Version != version {
-		return fmt.Errorf("version %d is not the last version for LP %s (party %d), last is %d", version, id, party, last.LiquidityProvision().Version)
-	}
-	if okRef {
-		if lRef := last.LiquidityProvision().Reference; ref != lRef {
-			return fmt.Errorf("expected reference %s, got %s (LP ID %s, party %s)", ref, lRef, id, party)
-		}
+		return fmt.Errorf("version %d is not the last version for LP %s (party %s), last is %d", version, id, party, last.LiquidityProvision().Version)
 	}
 	if amt != nil && last.LiquidityProvision().CommitmentAmount != amt.String() {
 		fmt.Errorf("commitment amount was %s, expected %s for last event for LP %s, party %s",
@@ -94,7 +88,6 @@ func parseLPEventTable(table *godog.Table) []RowWrapper {
 		"version",
 	}, []string{
 		"commitment amount",
-		"reference",
 		"final",
 	})
 }
@@ -116,14 +109,6 @@ func (l LPEventWrapper) CommitmentAmount() *num.Uint {
 		return nil
 	}
 	return l.row.MustUint("commitment amount")
-}
-
-func (l LPEventWrapper) Reference() (string, bool) {
-	if !l.row.HasColumn("reference") {
-		return "", false
-	}
-	r := l.row.MustStr("reference")
-	return r, true
 }
 
 func (l LPEventWrapper) Final() bool {